refactor(middleware): export middleware from utils and use it everywhere
